Add ability to cache found directories to file
For large directories Command-T will take a while to go through the
directories. This means that when opening vim and running <Leader>t the
first time it will take quite a while.

To work around this issue, add the feature to allow saving of the
directories found to a file on disk. Then when launching vim, Command-T
will load the directories cache directly from disk instead of going
through all the directories again.

Users can do :CommandTFlush, which will delete the saved file on disk,
if they wish to refresh the cache.