Allow setting which directories will cache to disk
In terms of editing regular files, a lot of the time the directory
restructure is small enough that command-t is responsive. Therefore,
there is really no need to cache to disk.

Therefore, adding a parameter to specify what directories specifically
need to cache to disk to allow command-t to be responsive enough on
first use.